+++ src/hooks.c
@@ -6,7 +6,7 @@
 /*   By: rcoetzer <rcoetzer@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/07/05 09:29:15 by rcoetzer          #+#    #+#             */
-/*   Updated: 2019/07/06 18:06:53 by rcoetzer         ###   ########.fr       */
+/*   Updated: 2019/07/07 01:26:49 by rcoetzer         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -86,11 +86,11 @@ int			ft_handelrotation(int key, t_env *env)
 	return (0);
 }
 
-double		ft_zoom_hndl(unsigned int y, unsigned int x)
+float		ft_zoom_hndl(unsigned int y, unsigned int x)
 {
-	double side;
-	double diag;
-	double ratio;
+	float side;
+	float diag;
+	float ratio;
 
 	side = 0;
 	if (y > x)
@@ -99,6 +99,6 @@ double		ft_zoom_hndl(unsigned int y, unsigned int x)
 		side = x;
 
 	diag = sqrt((pow(x , 2) + pow( y , 2)));
-	ratio = ((double)side / diag);
+	ratio = ((float)side / diag);
 	return ((side / (diag / (ratio / (side/2)))));
 }
